IntelliJ "inspect code" cleanup, and some reformatting

+++ modelstudio/src/com/hiveworkshop/rms/editor/model/Layer.java
@@ -389,7 +389,7 @@ public class Layer extends TimelineContainer implements Named {
 			final AnimFlag txFlag = find(MdlUtils.TOKEN_TEXTURE_ID);
 			for (int i = 0; i < txFlag.values.size(); i++) {
 				final Bitmap textureFoundFromDirtyId = ridiculouslyWrongTextureIDToTexture
-						.get((Integer) txFlag.values.get(i));
+						.get(txFlag.values.get(i));
 				final int newerTextureId = mdlr.getTextureId(textureFoundFromDirtyId);
 				txFlag.values.set(i, newerTextureId);
 				ridiculouslyWrongTextureIDToTexture.put(newerTextureId, textureFoundFromDirtyId);
